### 2022-09-09 #4749 @editorialbot
[PRE REVIEW]: MetaGenePipe: An Automated, Portable Pipeline for Contig-based Functional and Taxonomic Analysis
**Submitting author:** @ParkvilleData (<a href="http://orcid.org/0000-0002-7393-810X">Babak Shaban</a>)
**Repository:** https://github.com/ParkvilleData/MetaGenePipe/
**Branch with paper.md** (empty if default branch):
**Version:** v.1.0.0
