+++ docs/refs.bib
@@ -19,6 +19,7 @@
   pages    = "1861--1864",
   month    =  oct,
   year     =  2017,
+  doi      = "10.1002/jmv.24839",
   keywords = "Phlebovirus; Toscana virus; encephalitis; high throughput
               sequencing; travel",
   language = "en"
@@ -183,6 +184,7 @@
   pages     = "e17288",
   month     =  mar,
   year      =  2011,
+  doi       = "10.1371/journal.pone.0017288",
   language  = "en"
 }
 @ARTICLE{Hyatt2010-zh,
@@ -382,6 +384,7 @@
   volume    = "41 Suppl 1",
   pages     = "2643--2650",
   year      =  2012,
+  doi       = "10.3233/WOR-2012-0507-2643",
   language  = "en"
 }
 
@@ -414,6 +417,7 @@
   publisher = "IOS Press",
   volume    = "41 Suppl 1",
   pages     = "2656--2660",
+  doi       = "10.3233/wor-2012-0508-2656",
   year      =  2012,
   language  = "en"
 }
@@ -446,6 +450,7 @@
   volume    = "41 Suppl 1",
   pages     = "2661--2667",
   year      =  2012,
+  doi       = "10.3233/wor-2012-1032-2661",
   language  = "en"
 }
 
