### 2022-09-10 #4749 @jmschrei
Hi @rbturnbull @kthyng. I'm trying to assess whether this submission meets the requirement for substantial scholarly effort (**https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort**) and I'm having some difficulty, in part because I'm not as familiar with metagenomics as I am other fields of genomics. After reading the manuscript, the repo, and the associated documentation, I still don't know precisely what it is that this software does. I understand that there are pipelines that process data from metagenomic samples but, for example, I'm not sure what the statement "create an accurate taxonomic and functional characterization of the prokaryotic fraction of sequenced microbiomes." What is a "functional characterization" in this context? What is output as part of the "taxonomic classification"? What specific problems does this solve?

If you can revise the repo README and documentation to be more specific as to the high-level objectives of the package, as well as the specific steps one would use to achieve those goals, it'd be a lot easier for me to assess the scholarly effort. Keep in mind that the documentation should not be for a subject matter expert, but a user who is quickly trying to assess if this software will solve their problem. I'm not doubting the scientific merit of the software, but I don't fully understand from the current documentation what exactly it does or when one would use it.
