### 2022-09-12 #4749 @kthyng
@jmschrei Thanks for your comments. Given this, a scope query is appropriate to ping the editorial board and gather a consensus about the fit of this submission for JOSS. I'll start that process.

@rbturnbull It'd be helpful if you can address @jmschrei's comments in the meantime so we can consider that information alongside your submission. This process will take 1-2 weeks.
