### 2022-09-13 3a5866c
updated README and docs
