### 2022-09-13 368c30b
updated readme with tree and output definitions
