### 2022-09-13 #4749 @ParkvilleData
Hi @jmschrei, I have updated the README to include an output tree of all output files, descriptions of the output files and extra paragraphs detailing the usefulness of taxonomic profiling. There is also a table that represents how the profiles are represented in the outputs.

The workflow is a collection of "wrappers" that encompass bioinformatics standard software in the field of metagenomics. As stated in the paper, there are similar examples written in NextFlow (Nf-core and Muffin) and Snakemake (Atlas) but very limited options in WDL. The advantages of the workflow include it's versatility (in terms of substituting alignment databases) which can tailor the workflow to the researchers question of interest and the setup script can facilitate this by downloading required blast databases. WDL is a human readable workflow language and as such we have found has a much lower learning curve for lab based biologists, has a wealth of documentation, tutorials and examples, and doesn't need background knowledge in a programming language such as python.

Previous iterations of the workflow has been used by myself as a method in published papers to discover novel RNA viruses. The current iteration is the "polished" version which has been applied to the field of metagenomics.
