### 2022-09-15 #4749 @danielskatz
(just following up from the try by @rbturnbull a few days ago - editorialbot commands need to be the first thing in a comment)
