+++ docs/refs.bib
@@ -59,6 +59,7 @@
   publisher = "Elsevier BV",
   volume    =  175,
   number    =  2,
+  doi       = "10.1016/j.cell.2018.08.013",
   pages     = "530--543.e24",
   month     =  oct,
   year      =  2018,
@@ -606,6 +607,7 @@ url="https://doi.org/10.1007/978-1-4939-9173-0_6"
 
 @misc{krakenmanual,
   Title="Kraken taxonomic sequence classification system: Operating Manual", 
+  doi="10.1186/gb-2014-15-3-r46",
   url={https://ccb.jhu.edu/software/kraken/MANUAL.html}
 } 
 
@@ -802,6 +804,7 @@ url="https://doi.org/10.1007/978-1-4939-9173-0_6"
 @misc{voss_van_der_auwera_gentry_2022, 
   title={Full-stack genomics pipelining with GATK4 + WDL + Cromwell},
     url={https://f1000research.com/slides/6-1381},
+    doi="10.7490/f1000research.1114634.1",
     journal={F1000research.com},
     author={Voss, Kate and Van der Auwera, Geraldine and Gentry, Jeff},
     year={2022}
