### 2022-09-15 #4749 @ParkvilleData
Hi, I have updated Missing DOIs. Fastqc doesn't seem to have one.
