### 2022-09-15 #4749 @danielskatz
editorialbot commands need to be the first thing in a comment
