### 2022-10-13 #4851 @editorialbot
[REVIEW]: MetaGenePipe: An Automated, Portable Pipeline for Contig-based Functional and Taxonomic Analysis
**Submitting author:** @ParkvilleData (<a href="http://orcid.org/0000-0002-7393-810X">Babak Shaban</a>)
**Repository:** https://github.com/ParkvilleData/MetaGenePipe/
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.5
**Archive:** 10.26188/22032425
