### 2022-10-26 #4851 @Ebedthan
**Summary and overall critics**
We thank the authors for the proposed article which is a valuable study proposing a useful tool for bioinformaticians and computational biologists. I have particularly loved the idea and implementation that take into account the possibility to run it on servers. I have also appreciated the software website. However, this article and the associated code repository suffer from some correctable shortcomings.

**Major points**
* Statement of need: I would suggest the statement of need be shortened. The full statement could then be found on the software website or in the article.
* Installation instructions: Even if the software has a website that contains all the information about the installation of the software, I believe an installation section in the repository README is needed. I also believe that this would make it easier for an already convinced user to quickly install the software for use.
* Installation instructions: Java and Singularity which are required dependencies for MGP do not have a minimum supported version specified.
* Example usage: As for the installation instruction, the repo README should contain an example usage section.
* Community guidelines: The software community guideline should benefit from a section in the README repository.

**Minor points**
* Line 13: in its default form -> in its default format
* Line 14: which uses -> that uses
* Line 16: is freely available and is distributed -> is freely available and distributed
* Line 22: which downloads -> that downloads
* Line 26: infrastructure -> infrastructures
* Line 33: A missing comma: Currently -> Currently,
* Line 34: function -> functions
* Line 50: best practice -> best practices
* Line 52: Unnecessary comma: Singularity containers, and -> Singularity containers and
* Line 52 - 53: increases flexibility -> increases the flexibility
* Line 78: to facilitate co-assembly -> to facilitate the co-assembly
* Line 102: produces an fa file -> produces a Fasta file
* Line 102: aminoacid -> amino acid
* Line 113: SPARTAN high performance -> SPARTAN high-performance
* Add the software website link in the repository about section to ease its discovery by the user
