### 2022-10-26 bbbe1fd
Fix minor points from reviewer
