### 2022-10-27 1dcd717
:abc: updating nf-core/mag reference from preprint to published article
