+++ docs/refs.bib
@@ -730,11 +730,20 @@ url="https://doi.org/10.1007/978-1-4939-9173-0_6"
               }
 
 
-@article{krakau_straub_gourlé_gabernet_nahnsen_2021,
-    title={nf-core/mag: a best-practice pipeline for metagenome hybrid assembly and binning},
-    DOI={10.1101/2021.08.29.458094},
-    author={Krakau, Sabrina and Straub, Daniel and Gourlé, Hadrien and Gabernet, Gisela and Nahnsen, Sven},
-    year={2021}
+@article{krakau_straub_gourle_gabernet_nahnsen_2022,
+    author = {Krakau, Sabrina and Straub, Daniel and Gourlé, Hadrien and Gabernet, Gisela and Nahnsen, Sven},
+    title = "{nf-core/mag: a best-practice pipeline for metagenome hybrid assembly and binning}",
+    journal = {NAR Genomics and Bioinformatics},
+    volume = {4},
+    number = {1},
+    year = {2022},
+    month = {02},
+    abstract = "{The analysis of shotgun metagenomic data provides valuable insights into microbial communities, while allowing resolution at individual genome level. In absence of complete reference genomes, this requires the reconstruction of metagenome assembled genomes (MAGs) from sequencing reads. We present the nf-core/mag pipeline for metagenome assembly, binning and taxonomic classification. It can optionally combine short and long reads to increase assembly continuity and utilize sample-wise group-information for co-assembly and genome binning. The pipeline is easy to install-all dependencies are provided within containers-portable and reproducible. It is written in Nextflow and developed as part of the nf-core initiative for best-practice pipeline development. All codes are hosted on GitHub under the nf-core organization https://github.com/nf-core/mag and released under the MIT license.}",
+    issn = {2631-9268},
+    doi = {10.1093/nargab/lqac007},
+    url = {https://doi.org/10.1093/nargab/lqac007},
+    note = {lqac007},
+    eprint = {https://academic.oup.com/nargab/article-pdf/4/1/lqac007/42366621/lqac007.pdf},
 }
 
 @article{kieser_brown_zdobnov_trajkovski_mccue_2020,
