### 2022-10-27 d04d3a4
:abc: updating README to include instructions for installation, usage, community guidelines and a link to the documentation (openjournals/joss-reviews#4851)
