+++ paper.md
@@ -43,7 +43,7 @@ bibliography: docs/refs.bib
 MetaGenePipe (MGP) is an efficient, flexible, portable, and scalable metagenomics pipeline that uses performant bioinformatics software suites and genomic databases to create an accurate taxonomic and functional characterization of the prokaryotic fraction of sequenced microbiomes. Written in the Workflow Definition Language (WDL), MGP produces output that is useful in its default format, or that can be used for further downstream analysis. MGP is a pipeline-development best practice tool that uses Singularity [@kurtzer_sochat_bauer_2017] for containerization and includes a setup script that downloads the necessary databases for setup. The source code for MGP is freely available and distributed under the [Apache 2.0 license](https://www.apache.org/licenses/LICENSE-2.0).
 
 
-There are currently several assembly-based taxonomic software suites such as MG-RAST [@keegan_glass_meyer_2016], MMseqs2 [@10.1093/bioinformatics/btab184], and a few that make use of a workflow language, including nf-core and Muffin written in Nextflow, and Atlas written in Snakemake [@krakau_straub_gourlé_gabernet_nahnsen_2021; @di_tommaso_chatzou_floden_barja_palumbo_notredame_2017; @van_damme_hölzer_viehweger_müller_bongcam-rudloff_brandt_2021;  @kieser_brown_zdobnov_trajkovski_mccue_2020; @mölder_jablonski_letcher_hall_tomkins-tinch_sochat_forster_lee_twardziok_kanitz_et_al._2021]. The main advantage of MGP over MG-RAST is its ease of installation on local infrastructure, as it only requires running a setup script that downloads a supplied Singularity image from SylabsCloud, the latest version of Cromwell [@voss_van_der_auwera_gentry_2022], Koalafam HMMER profiles [@aramaki_blanc-mathieu_endo_ohkubo_kanehisa_goto_ogata_2019], and the Swiss-Prot database [@uniprot_consortium_2018] which is converted to the DIAMOND aligner [@Buchfink2015-rn] format. This setup allows MGP to be used on a range of computing infrastructures across institutions. 
+There are currently several assembly-based taxonomic software suites such as MG-RAST [@keegan_glass_meyer_2016], MMseqs2 [@10.1093/bioinformatics/btab184], and a few that make use of a workflow language, including nf-core and Muffin written in Nextflow, and Atlas written in Snakemake [@krakau_straub_gourle_gabernet_nahnsen_2022; @di_tommaso_chatzou_floden_barja_palumbo_notredame_2017; @van_damme_hölzer_viehweger_müller_bongcam-rudloff_brandt_2021;  @kieser_brown_zdobnov_trajkovski_mccue_2020; @mölder_jablonski_letcher_hall_tomkins-tinch_sochat_forster_lee_twardziok_kanitz_et_al._2021]. The main advantage of MGP over MG-RAST is its ease of installation on local infrastructure, as it only requires running a setup script that downloads a supplied Singularity image from SylabsCloud, the latest version of Cromwell [@voss_van_der_auwera_gentry_2022], Koalafam HMMER profiles [@aramaki_blanc-mathieu_endo_ohkubo_kanehisa_goto_ogata_2019], and the Swiss-Prot database [@uniprot_consortium_2018] which is converted to the DIAMOND aligner [@Buchfink2015-rn] format. This setup allows MGP to be used on a range of computing infrastructures across institutions. 
 
  
 
