+++ preprint.tex
@@ -11,7 +11,7 @@ analysis. There are currently several contig-based taxonomic software
 suites such as MG-RAST \autocite{keegan_glass_meyer_2016}, MMseqs2
 \autocite{10.1093/bioinformatics/btab184}, and a few which make use of a
 workflow language such as nf-core
-\autocite{krakau_straub_gourlé_gabernet_nahnsen_2021} (written in
+\autocite{krakau_straub_gourle_gabernet_nahnsen_2022} (written in
 Nextflow
 \autocite{di_tommaso_chatzou_floden_barja_palumbo_notredame_2017}),
 Muffin

[…]
