### 2022-10-27 1a5f67e
:abc: moving info about fastq files to usage section in readme
