### 2022-10-27 e38dd5b
:abc: adding minimum version numbers for java and singularity (openjournals/joss-reviews#4851)
