### 2022-10-28 6b47384
Shorten statement of need, better align with example
