### 2022-11-03 c3b7258
Update link to singularity image
