### 2022-11-04 4cf5ac6
Update name of singularity file and create kegg directory
