### 2022-11-06 7d8ceb8
Create kegg folder for kegg_all file if it doesn't exist
