### 2022-12-04 56936f7
Fix hmmer taxon task to generate OTU file
