### 2022-12-05 d0393dc
Start addressing reviewer 2's comments
