### 2022-12-22 553f170
Output section in README and rename brite levels
