+++ README.rst
@@ -126,13 +126,18 @@ There are five main folders of output (Assembly, Gene Prediction, Read Alignment
 
 An example tree of the the output directory and the associated output definitions are below:
 
+Quality Control
+
+* {sampleName}.TG_R1_fastqc.zip: Fastqc output for each of the individual sample files
+* multiqc_report.html: Combined report of all fastqc files
+
 The Assembly directory contains the following
 
-* merged.contigs.k27.fa: Kmer assembled contigs: assembled contigs for the kmer values, represented in the "intermediate_contigs" folder
-* merged.megahit.contigs.fa: Final assembled contigs
-* merged.37.fastg: A fastg file. Fastg is the assembly graph produced by the assembler.
-* merged.megahit.blast.out: Raw blast results for the contigs
-* merged.megahit.blast.parsed: Blast results parsed to be easily viewed in tsv format
+* {sampleName}.megahit.contigs.fa: Final assembled contigs
+* {sampleName}.{kmer}.fastg: Assembly graph for {kmer} assembled contigs, where {kmer} produces the largest assembled contig file size in the `intermediate_contigs` folder
+* intermediate_contigs: a folder containing all intermediate assembled contigs {sampleName}.contigs.k{kmer}.fastg
+* {sampleName}.megahit.blast.out: Raw blast results for the contigs
+* {sampleName}.megahit.blast.parsed: Blast results parsed to be easily viewed in tsv format
 
 Gene prediction contains the output from prodigal
 
@@ -143,11 +148,6 @@ Gene prediction contains the output from prodigal
 * Merge.prodigal.potential_genes.fa: Prodigal starts file
 * Merge.prodigal.proteins.fa: Predicted gene amino acid sequences
 * Merge.xml: XML output of alignment of predicted Amino Acids to NCBI database (We chose swissprot, but any blast database can be substituted)
-
-Quality Control
-
-* SRR5808831.TG_R1_fastqc.zip: Fastqc output for each of the individual sample files
-* multiqc_report.html: Combined report of all fastqc files
 
 Read Alignment
 
@@ -168,57 +168,61 @@ Output Tree
 
    .
    ├── assembly
-   │   ├── intermediate_contigs
-   │   │   ├── merged.contigs.k27.fa
-   │   │   ├── merged.contigs.k37.fa
-   │   │   ├── merged.contigs.k47.fa
-   │   │   ├── merged.contigs.k57.fa
-   │   │   ├── merged.contigs.k67.fa
-   │   │   ├── merged.contigs.k77.fa
-   │   │   ├── merged.contigs.k87.fa
-   │   │   └── merged.contigs.k97.fa
-   │   ├── merged.37.fastg
-   │   ├── merged.megahit.blast.out
-   │   ├── merged.megahit.blast.parsed
-   │   └── merged.megahit.contigs.fa
+   │   ├── combined.57.fastg
+   │   ├── combined.megahit.blast.out
+   │   ├── combined.megahit.blast.parsed
+   │   ├── combined.megahit.contigs.fa
+   │   └── intermediate_contigs
+   │       ├── combined.contigs.k27.fa
+   │       ├── combined.contigs.k37.fa
+   │       ├── combined.contigs.k47.fa
[…]
