### 2022-12-22 9a345ee
:abc: adding more info to output section of paper
