+++ docs/refs.bib
@@ -910,3 +910,41 @@ url="https://doi.org/10.1007/978-1-4939-9173-0_6"
     issn = "2045-2322",
     da = "2020/07/01"
 }
+
+@article{kang2019,
+    author = "Kang, Dongwan D and Li, Feng and Kirton, Edward and Thomas, Ashleigh and Egan, Rob and An, Hong and Wang, Zhong",
+    type = "Journal Article",
+    title = "MetaBAT 2: an adaptive binning algorithm for robust and efficient genomereconstruction from metagenome assemblies.",
+    journal = "PeerJ",
+    date = "2019",
+    volume = "7",
+    pages = "e7359",
+    keywords = "Clustering | Metagenome binning | Metagenomics",
+    issn = "2167-8359 (Print); 2167-8359 (Electronic); 2167-8359 (Linking)",
+    abstract = "We previously reported on MetaBAT, an automated metagenome binning software toolto reconstruct single genomes from microbial communities for subsequent analysesof uncultivated microbial species. MetaBAT has become one of the most popularbinning tools largely due to its computational efficiency and ease of use,especially in binning experiments with a large number of samples and a largeassembly. MetaBAT requires users to choose parameters to fine-tune itssensitivity and specificity. If those parameters are not chosen properly, binningaccuracy can suffer, especially on assemblies of poor quality. Here, we developedMetaBAT 2 to overcome this problem. MetaBAT 2 uses a new adaptive binningalgorithm to eliminate manual parameter tuning. We also performed extensivesoftware engineering optimization to increase both computational and memoryefficiency. Comparing MetaBAT 2 to alternative software tools on over 100 realworld metagenome assemblies shows superior accuracy and computing speed. Binninga typical metagenome assembly takes only a few minutes on a single commodityworkstation. We therefore recommend the community adopts MetaBAT 2 for theirmetagenome binning experiments. MetaBAT 2 is open source software and availableat https://bitbucket.org/berkeleylab/metabat.",
+    year = "2019",
+    doi = "10.7717/peerj.7359",
+    pmid = "31388474",
+    own = "NLM",
+    stat = "PubMed-not-MEDLINE",
+    lr = "20201001",
+    lid = "10.7717/peerj.7359 [doi]; e7359",
+    au = "Kang DD; Li F; Kirton E; Thomas A; Egan R; An H; Wang Z",
+    ad = "Department of Energy, Joint Genome Institute, Walnut Creek, CA, USA.; School of Computer Science and Technology, University of Science and Technologyof China, Hefei, Anhui, China.; Department of Energy, Joint Genome Institute, Walnut Creek, CA, USA.; Department of Energy, Joint Genome Institute, Walnut Creek, CA, USA.; Department of Energy, Joint Genome Institute, Walnut Creek, CA, USA.; School of Computer Science and Technology, University of Science and Technologyof China, Hefei, Anhui, China.; Department of Energy, Joint Genome Institute, Walnut Creek, CA, USA.; Environmental Genomics and System Biology Division, Lawrence Berkeley NationalLaboratory, Berkeley, CA, USA.; School of Natural Sciences, University of California at Merced, Merced, CA, USA.",
+    auid = "ORCID: 0000-0003-1974-1768; ORCID: 0000-0001-7422-5649; ORCID: 0000-0001-7080-7801; ORCID: 0000-0002-6307-0458",
+    la = "eng",
+    dep = "20190726",
+    pl = "United States",
+    ta = "PeerJ",
+    jid = "101603425",
+    pmc = "PMC6662567",
[…]
