+++ README.rst
@@ -122,44 +122,51 @@ Read the `documentation <https://parkvilledata.github.io/MetaGenePipe/usage.html
 Output
 ======
 
-There are five main folders of output (Assembly, Gene Prediction, Read Alignment, QC and Taxon) and one intermediary (data) which contains the merging of raw samples and the output from Trim Galore. 
+There are four main output folders: qc (quality control), assembly, readalignment, and geneprediction and one intermediary, data, which contains the samples for assembly after running through TrimGalore and concatenating the samples for co-assembly if specified. 
 
 An example tree of the the output directory and the associated output definitions are below:
 
 Quality Control
-
-* {sampleName}.TG_R1_fastqc.zip: Fastqc output for each of the individual sample files
+* trimmed
+  * {sampleName}.T{G|T}_R{1|2}.fq.gz: Trimmed output for each of the individual sample files, TG if the chosen trimmer is TrimGalore, and TT if it is Trimmomatic
+* fastqc
+  * {sampleName}.T{G|T}_R{1|2}_fastqc.zip: Fastqc output for each of the individual sample files
 * multiqc_report.html: Combined report of all fastqc files
+* flash
+  * {sampleName}.extendedFrags.fastq: [ADD HERE]
 
-The Assembly directory contains the following
+Data
+* {sampleName}_R{1|2}.fq.gz Sample files after trimming and/or concatenating for co-assembly. If files are concatenated for co-assembly, the sample name is set to be `combined`
 
+Assembly
 * {sampleName}.megahit.contigs.fa: Final assembled contigs
 * {sampleName}.{kmer}.fastg: Assembly graph for {kmer} assembled contigs, where {kmer} produces the largest assembled contig file size in the `intermediate_contigs` folder
 * intermediate_contigs: a folder containing all intermediate assembled contigs {sampleName}.contigs.k{kmer}.fastg
 * {sampleName}.megahit.blast.out: Raw blast results for the contigs
 * {sampleName}.megahit.blast.parsed: Blast results parsed to be easily viewed in tsv format
 
-Gene prediction contains the output from prodigal
+Read alignment
+* {sampleName}.T{G|T}.flagstat.txt: Samtools flagstat output. Reports statistics on alignment of reads back to assembled contigs
+* {sampleName}.T{G|T}.sam: Alignment of reads back to contigs in SAM format
+* {sampleName}.T{G|T}.sorted.bam: Alignment of reads back to contigs in BAM format
 
-* Merge.hmmer.out: Raw hmmer output aligned to Koalafam profiles
-* Merge.hmmer.tblout: Parsed hmmer output aligned to Koalafam profiles
-* Merge.prodigal.genes.fa: Gene coordinates file (Genbank like file)
-* Merge.prodigal.nucl.genes.fa: Predicted gene nucleotide sequences
-* Merge.prodigal.potential_genes.fa: Prodigal starts file
-* Merge.prodigal.proteins.fa: Predicted gene amino acid sequences
-* Merge.xml: XML output of alignment of predicted Amino Acids to NCBI database (We chose swissprot, but any blast database can be substituted)
-
-Read Alignment
-
-* SRR5808831.TG.flagstat.txt: Samtools flagstat output. Reports statistics on alignment of reads back to assembled contigs
-* SRR5808831.TG.sam: Alignment of reads back to contigs in SAM format
-* SRR5808831.TG.sorted.bam: Alignment of reads back to contigs in BAM format
-
-Taxon output
-
[…]
