+++ README.rst
@@ -128,6 +128,7 @@ There are four main output folders: qc (quality control), assembly, readalignmen
 An example tree of the the output directory and the associated output definitions are below:
 
 Quality Control
+
 * trimmed
   * {sampleName}.T{G|T}_R{1|2}.fq.gz: Trimmed output for each of the individual sample files, TG if the chosen trimmer is TrimGalore, and TT if it is Trimmomatic
 * fastqc
@@ -137,9 +138,11 @@ Quality Control
   * {sampleName}.extendedFrags.fastq: [ADD HERE]
 
 Data
+
 * {sampleName}_R{1|2}.fq.gz Sample files after trimming and/or concatenating for co-assembly. If files are concatenated for co-assembly, the sample name is set to be `combined`
 
 Assembly
+
 * {sampleName}.megahit.contigs.fa: Final assembled contigs
 * {sampleName}.{kmer}.fastg: Assembly graph for {kmer} assembled contigs, where {kmer} produces the largest assembled contig file size in the `intermediate_contigs` folder
 * intermediate_contigs: a folder containing all intermediate assembled contigs {sampleName}.contigs.k{kmer}.fastg
@@ -147,11 +150,13 @@ Assembly
 * {sampleName}.megahit.blast.parsed: Blast results parsed to be easily viewed in tsv format
 
 Read alignment
+
 * {sampleName}.T{G|T}.flagstat.txt: Samtools flagstat output. Reports statistics on alignment of reads back to assembled contigs
 * {sampleName}.T{G|T}.sam: Alignment of reads back to contigs in SAM format
 * {sampleName}.T{G|T}.sorted.bam: Alignment of reads back to contigs in BAM format
 
 Gene prediction
+
 * {sampleName}.megahit.proteins.fa.xml.out.xml: XML output of alignment of predicted Amino Acids to NCBI database (We chose swissprot, but any blast database can be substituted)
 * diamond
   * {sampleName}.megahit.proteins.fa.xml.out:
