### 2022-12-22 36c145d
:abc: updates to outputs section of paper and readme
