+++ docs/refs.bib
@@ -863,17 +863,6 @@ url="https://doi.org/10.1007/978-1-4939-9173-0_6"
     pages={1420-1428}
 }
 
-@article{li_liu_luo_sadakane_lam_2015,
-    title={MEGAHIT: an ultra-fast single-node solution for large and complex metagenomics assembly via succinct de Bruijn graph},
-    volume={31},
-    DOI={10.1093/bioinformatics/btv033},
-    number={10},
-    journal={Bioinformatics},
-    author={Li, Dinghua and Liu, Chi-Man and Luo, Ruibang and Sadakane, Kunihiko and Lam, Tak-Wah},
-    year={2015},
-    pages={1674-1676}
-}
-
 @article{hyatt_chen_locascio_land_larimer_hauser_2010,
     title={Prodigal: prokaryotic gene recognition and translation initiation site identification},
     volume={11},
@@ -948,3 +937,48 @@ url="https://doi.org/10.1007/978-1-4939-9173-0_6"
     so = "PeerJ. 2019 Jul 26;7:e7359. doi: 10.7717/peerj.7359. eCollection 2019.",
     aid = "7359 [pii]"
 }
+@article {Gonzalez-Tortuero2021.12.11.472104,
+	author = {Gonz{\'a}lez-Tortuero, Enrique and Krishnamurthi, Revathy and Allison, Heather E. and Goodhead, Ian B. and James, Chlo{\"e} E.},
+	title = {Comparative analysis of gene prediction tools for viral genome annotation},
+	elocation-id = {2021.12.11.472104},
+	year = {2021},
+	doi = {10.1101/2021.12.11.472104},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {The number of newly available viral genomes and metagenomes has increased exponentially since the development of high throughput sequencing platforms and genome analysis tools. Bioinformatic annotation pipelines are largely based on open reading frame (ORF) calling software, which identifies genes independently of the sequence taxonomical background. Although ORF-calling programs provide a rapid genome annotation, they can misidentify ORFs and start codons; errors that might be perpetuated and propagated over time. This study evaluated the performance of multiple ORF-calling programs for viral genome annotation against the complete RefSeq viral database. Programs outputs varied when considering the viral nucleic acid type versus the viral host. According to the number of ORFs, Prodigal and Metaprodigal were the most accurate programs for DNA viruses, while FragGeneScan and Prodigal generated the most accurate outputs for RNA viruses. Similarly, Prodigal outperformed the benchmark for viruses infecting prokaryotes, and GLIMMER and GeneMarkS produced the most accurate annotations for viruses infecting eukaryotes. When the coordinates of the ORFs were considered, Prodigal scored high for all scenarios except for RNA viruses, where GeneMarkS generated the most reliable results. Overall, the quality of the coordinates predicted for RNA viruses was poorer than for DNA viruses, suggesting the need for improved ORF-calling programs to deal with RNA viruses. Moreover, none of the ORF-calling programs reached 90\% accuracy for annotation of DNA viruses. Any automatic annotation can still be improved by manual curation, especially when the presence of ORFs is validated with wet-lab experiments. However, our evaluation of the current ORF-calling programs is expected to be useful for the improvement of viral genome annotation pipelines and highlights the need for more expression data to improve the rigor of reference genomes.Competing Interest StatementThe authors have declared no competing interest.},
+	URL = {https://www.biorxiv.org/content/early/2021/12/13/2021.12.11.472104},
+	eprint = {https://www.biorxiv.org/content/early/2021/12/13/2021.12.11.472104.full.pdf},
+	journal = {bioRxiv}
+}
+
[…]

[…]
