+++ paper.md
@@ -94,7 +94,7 @@ The protein coding sequences are also aligned to the [KoalaFam HMM profiles](htt
 
 ## Outputs and interpretation
 
-The principal outputs for downstream analysis are the assembly, the mapped reads, and tables of the functional and taxonomic (organisms) counts of the genes found in the contigs. For the assembly, both the final contigs as well as the assembly graph which users can visualize [@10.1093/bioinformatics/btv383] or use for downstream graph-based metagenome binning [@10.1007/978-3-031-04749-7_5]. The mapping results in SAM/BAM mapping files for each pair of read files, which can be used for downstream metagenome binning applications, or users could use these to obtain a list of contigs with read depth metrics by running these through the [jgi_summarize_bam_contig_depths](https://bitbucket.org/berkeleylab/metabat/src/master/src/jgi_summarize_bam_contig_depths.cpp) tool, available through MetaBAT [@kang2019]. The taxonomic and functional tables of gene counts produced in the gene prediction sub-workflow are suitable for downstream statistical analysis. Several other potentially useful intermediate files are saved in each of the sub-workflows. A description of each of the output files and an example hierarchical structure of the output directory are available in the [MGP documentation](https://parkvilledata.github.io/MetaGenePipe/workflow.html#output).
+The principal outputs for downstream analysis are the assembly, the mapped reads, and tables of the functional and taxonomic (organisms) counts of the genes found in the contigs. For the assembly, outputs include the final contigs as well as the assembly graph which users can visualize [@10.1093/bioinformatics/btv383] or use for downstream graph-based metagenome binning [@10.1007/978-3-031-04749-7_5]. The mapping results in SAM/BAM mapping files for each pair of read files, which can be used for downstream metagenome binning applications, or users could use these to obtain a list of contigs with read depth metrics by running these through the [jgi_summarize_bam_contig_depths](https://bitbucket.org/berkeleylab/metabat/src/master/src/jgi_summarize_bam_contig_depths.cpp) tool, available through MetaBAT [@kang2019]. The taxonomic and functional tables of gene counts produced in the gene prediction sub-workflow are suitable for downstream statistical analysis. Several other potentially useful intermediate files are saved in each of the sub-workflows. A description of each of the output files and an example hierarchical structure of the output directory are available in the [MGP documentation](https://parkvilledata.github.io/MetaGenePipe/workflow.html#output).
 
 
 ## Resource usage and infrastructure requirements 
