### 2023-02-07 b67f8a6
:abc: reducing the significant figures in the Resource usage and infrastructure requirements section
