+++ paper.md
@@ -99,7 +99,7 @@ The principal outputs for downstream analysis are the assembly, the mapped reads
 
 ## Resource usage and infrastructure requirements 
 
-MGP uses Unix’s `time` tool to measure the resources used by each task, such as CPU usage, file size, elapsed time, and system time. This output can be visualized and used to inform resource requests when using a job scheduler on high-performance computing infrastructure. Table 1 shows indicative resource usage for processing paired-end samples of 25,000 reads each run on the University of Melbourne SPARTAN high-performance computing system consisting of Intel Xeon Gold 6154 3GHz CPUs. Running Cromwell on the head node took 2 minutes and 22.6 seconds (excluding time spent on the queue) and required a maximum memory of 837168 kbytes.
+MGP uses Unix’s `time` tool to measure the resources used by each task, such as CPU usage, file size, elapsed time, and system time. This output can be visualized and used to inform resource requests when using a job scheduler on high-performance computing infrastructure. Table 1 shows indicative resource usage for processing paired-end samples of 25,000 reads each run on the University of Melbourne SPARTAN high-performance computing system consisting of Intel Xeon Gold 6154 3GHz CPUs. Running Cromwell on the head node took 2 minutes and 23 seconds and required a maximum memory of 840 MB.
 
 MGP can be run locally on a laptop, a virtual machine, or in a high-performance computing setting. 
 
