### 2023-02-07 #4851 @rbturnbull
Hi @jmschrei - we've released a version of the software and added it to Figshare: https://melbourne.figshare.com/articles/software/MetaGenePipe_An_Automated_Portable_Pipeline_for_Contig-based_Functional_and_Taxonomic_Analysis/22032425

Here is the DOI: 10.26188/22032425.v1 (https://doi.org/10.26188/22032425.v1)

I also made a small change to the paper, rounding the report of the time and memory usage in the 'Resource usage and infrastructure requirements' section so there weren't so many irrelevant significant figures.
